Make bind work in a more comprehensive way
Binding on a success will result in an array with both the new result
and the old result in it. Binding on a failure will result in the same
failure.

It's not completely right: binding on a `Success a` with a function
`a -> b` should result in a `Success b`. Here's some more work to do.